fix: 체크버튼 쿼리 문제 해결
